requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
cloudipsp==1.0.2
currencies==2020.12.12
Django==4.0.6
django-extensions==3.2.0
django-filter==22.1
django-modeltranslation==0.18.2
django-phonenumber-field==6.3.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.7
phonenumbers==8.12.51
Pillow==9.2.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
webencodings==0.5.1
